Move company.go to service.companies
